requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
python-docx==0.8.11
Werkzeug==2.0.3
Jinja2==3.0.3
Flask-WTF==1.0.1
WTForms==3.0.0
SQLAlchemy==1.4.49
